+++ packages/widgetbook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **FEAT**: Add [Semantics Addon](https://docs.widgetbook.io/addons/semantics-addon). ([#1402](https://github.com/widgetbook/widgetbook/pull/1402))
 - **FEAT**: Add a banner to the bottom of the navigation panel that shows the total number of components and use-cases. ([#1405](https://github.com/widgetbook/widgetbook/pull/1405))
+- **FEAT**: Add [`home` widget](https://docs.widgetbook.io/guides/customization#home-widget) to customize the startup screen. ([#1418](https://github.com/widgetbook/widgetbook/pull/1418))
 - **REFACTOR**: Allow [`inspector`](https://pub.dev/packages/inspector) v3. ([#1407](https://github.com/widgetbook/widgetbook/pull/1407))
 - **FIX**: Remove `ExcludeSemantics` from Widgetbook UI; to allow screen readers. ([#1401](https://github.com/widgetbook/widgetbook/pull/1401) - by [@Goddchen](https://github.com/Goddchen))
 - **REFACTOR**: Replace [`device_frame`](https://pub.dev/packages/device_frame) with [`device_frame_plus`](https://pub.dev/packages/device_frame_plus). ([#1411](https://github.com/widgetbook/widgetbook/pull/1411))